# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Payproxy API",
        "description": "API pública para emissão de boletos bancários com split de pagamento. Autenticação via API Key no header `X-Api-Key`.",
        "contact": {
            "name": "Ciberian",
            "email": "suporte@ciberian.com.br"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Produção"
        }
    ],
    "paths": {
        "/boletos/batch": {
            "post": {
                "tags": [
                    "Boletos"
                ],
                "summary": "Emitir lote de boletos",
                "description": "Cria um lote de até 500 boletos. O processamento é assíncrono: retorna HTTP 202 com o ID do lote para consulta posterior. A operação é idempotente por `external_ref` por tenant.",
                "operationId": "9bb1004f9d2cae05c79da89d1f13be89",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "external_ref",
                                    "boletos"
                                ],
                                "properties": {
                                    "external_ref": {
                                        "description": "Referência única do lote por tenant",
                                        "type": "string",
                                        "example": "LOTE-IPTU-2026-07"
                                    },
                                    "boletos": {
                                        "type": "array",
                                        "items": {
                                            "type": "object"
                                        },
                                        "maxItems": 500
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Lote aceito para processamento"
                    },
                    "401": {
                        "description": "API Key inválida ou ausente"
                    },
                    "409": {
                        "description": "external_ref já existe para este tenant"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "ApiKey": []
                    }
                ]
            }
        },
        "/boletos/batch/{batch_id}": {
            "get": {
                "tags": [
                    "Boletos"
                ],
                "summary": "Consultar status do lote",
                "description": "Retorna o status de processamento do lote e os resultados individuais de cada boleto.",
                "operationId": "0dfe1850823df86e708a841b519004fa",
                "parameters": [
                    {
                        "name": "batch_id",
                        "in": "path",
                        "description": "ID do lote",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status do lote"
                    },
                    "404": {
                        "description": "Lote não encontrado"
                    }
                },
                "security": [
                    {
                        "ApiKey": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "ApiKey": {
                "type": "apiKey",
                "description": "API Key gerada no backoffice Payproxy. Formato: `ppx_<40 chars>`.",
                "name": "X-Api-Key",
                "in": "header"
            }
        }
    },
...
}